HSEARCH-3107 Fix Jenkins pipelines
