+++ jenkins/performance-search5-lucene.groovy
@@ -34,6 +34,7 @@ pipeline {
             steps {
                 unstash name: 'jar'
                 sh 'docker run --name postgresql -p 5431:5432 -e POSTGRES_USER=username -e POSTGRES_PASSWORD=password -e POSTGRES_DB=database -d postgres:10.5'
+                sleep(time:10,unit:"SECONDS") // wait for postgres to be ready
                 sh 'mkdir -p output'
                 sh """ \
 					java \
